WebApi/Application/UserOperations/Command/CreateUser/CreateUserCommand.cs
using AutoMapper;
using WebApi.DbOperations;
using WebApi.Entities;

namespace WebApi.Application.UserOperations.Command.CreateUser
{
    public class CreateUserCommand
    {
        public CreateUserModel Model { get; set; }
        private readonly IBookStoreDbContext _db;
        private readonly IMapper _mapper;

        public CreateUserCommand(IBookStoreDbContext db, IMapper mapper)
        {
            _db = db;
            _mapper = mapper;
        }

        public void Handler()
        {

            User user = _db.Users.SingleOrDefault(x => x.Email == Model.Email);
            if (user is not null)
                throw new InvalidOperationException("Kulanıcı Zaten Mevcut");

            user = _mapper.Map<User>(Model);

            _db.Users.Add(user);
            _db.SaveChanges();
        }

    }

    public class CreateUserModel
    {
        public string Name { get; set; }

        public string Surname { get; set; }

        public string Email { get; set; }

        public string Password { get; set; }

    }
}
